requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
fake-useragent==0.1.11
Flask==1.1.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.5.1
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
soupsieve==2.2
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
